fig6a/resnet/requirements.txt
absl_py==1.0.0
tensorflow_model_optimization==0.7.2
